requirements.txt
weaviate-client==3.25.3
sentence-transformers>=2.7.0
torch>=1.9.0
transformers>=4.21.0
numpy>=1.21.0 